Added file for the git lecture part 1
The material describes what we will cover in the first
of two lectures featuring git. Since the whole point is to
use git and familiarize the process, the material itself should
be hosted here.

The document and the repository's licence is undefined.
Whether there will be additional material on this repository
following this effort is also unclear.